+++ packages/help/index.js
@@ -1,0 +1,11 @@
+import HelpProvider, {
+  Help,
+  useHelp,
+  constants,
+  triggerFieldHelp,
+  FieldHelpIcon,
+} from './Help';
+
+export { Help, useHelp, constants, triggerFieldHelp, FieldHelpIcon };
+
+export default HelpProvider;