+++ packages/hooks/src/useCurrentUser.js
@@ -2,4 +2,5 @@ import { avUserApi } from '@availity/api-axios';
 import { useQuery } from 'react-query';
 
 const fetchUser = async () => avUserApi.me();
-export default () => useQuery(['user'], fetchUser);
+
+export default options => useQuery(['user'], fetchUser, options);