+++ packages/select/src/Select.js
@@ -7,6 +7,7 @@ import Creatable from 'react-select/creatable';
 import Async from 'react-select-async-paginate';
 import get from 'lodash.get';
 import isFunction from 'lodash.isfunction';
+import isEqual from 'lodash.isequal';
 
 const {
   DownChevron,
@@ -35,6 +36,27 @@ const createOption = (label, labelKey = 'label', valueKey = 'value') => ({
   [valueKey]: label.toLowerCase().replace(/\W/g, ''),
 });
 
+const areValueAndOptionValueEqual = (value, optionValue) => {
+  return isEqual(value, optionValue);
+};
+
+const selectAllOption = {
+  label: 'Select all',
+  value: '*',
+};
+
+const validateSelectAllOptions = options => {
+  const filtered = options.filter(
+    option => option.value === selectAllOption.value
+  );
+  if (filtered.length > 0) {
+    // eslint-disable-next-line no-console
+    console.warn(
+      `An option contains the value: ${selectAllOption.value}. This value is used by the Select All option.`
+    );
+  }
+};
+
 const Select = ({
   name,
   validate,
@@ -46,6 +68,7 @@ const Select = ({
   onChange: onChangeCallback,
   autofill,
   creatable,
+  allowSelectAll,
   ...attributes
 }) => {
   const [
@@ -87,8 +110,8 @@ const Select = ({
 
   const findOptionFromValue = (value, options) =>
     Array.isArray(options) &&
-    [...options, ...newOptions].filter(
-      option => getOptionValue(option) === value
+    [...options, ...newOptions].filter(option =>
+      areValueAndOptionValueEqual(value, getOptionValue(option))
     )[0];
 
   const getViewValue = () => {
@@ -112,6 +135,13 @@ const Select = ({
   }
 
   const onChangeHandler = async newValue => {
+    if (
+      newValue.length > 0 &&
+      newValue[newValue.length - 1].value === selectAllOption.value
+    ) {
+      newValue = options;
+    }
+
     const newVal = prepValue(newValue);
     const isOverMax =
       maxLength &&
@@ -197,6 +227,29 @@ const Select = ({
     }
   };
 
+  let selectOptions;
+  if (!attributes.loadOptions) {
+    if (allowSelectAll && attributes.isMulti) {
+      if (
+        [...options, ...newOptions].length > 0 &&
+        (values[name] === undefined ||
+          values[name].length < [...options, ...newOptions].length)
+      ) {
+        validateSelectAllOptions([...options, ...newOptions]);
+        selectOptions = [selectAllOption, ...options, ...newOptions];
+      } else {
+        selectOptions = [...options, ...newOptions];
+      }
+    } else {
+      selectOptions = [...options, ...newOptions];
+    }
+  }
+
+  if (attributes.loadOptions && allowSelectAll) {
+    // eslint-disable-next-line no-console
+    console.warn('allowSelectAll is ignored when loadOptions is defined.');
+  }
+
   return (
     <Tag
       {...field}
@@ -220,9 +273,7 @@ const Select = ({
       getOptionValue={getOptionValue}
       closeMenuOnSelect={!attributes.isMulti}
       components={components}
-      options={
-        !attributes.loadOptions ? [...options, ...newOptions] : undefined
-      }
+      options={selectOptions}
       defaultOptions
       styles={{
         ...styles,
@@ -311,6 +362,7 @@ Select.propTypes = {
   onChange: PropTypes.func,
   creatable: PropTypes.bool,
   autofill: PropTypes.oneOfType([PropTypes.bool, PropTypes.object]),
+  allowSelectAll: PropTypes.bool,
 };
 
 export default Select;