+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.9.2 (2020-08-13)
+
+**Note:** Version bump only for package @availity/pagination
+
+
+
+
+
+## 2.9.1 (2020-06-29)
+
+**Note:** Version bump only for package @availity/pagination
+
+
+
+
+
 # 2.9.0 (2020-05-29)